agent/requirements.txt
azure-ai-agentserver-agentframework==1.0.0b17
python-dotenv>=1.0.1,<2.0.0